+++ MaxMind.GeoIP2.UnitTests/DatabaseReaderTests.cs
@@ -95,6 +95,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 Assert.False(response.IsPublicProxy);
                 Assert.False(response.IsTorExitNode);
                 Assert.Equal(ipAddress, response.IPAddress);
+                Assert.Equal("1.2.0.0/16", response.Network.ToString());
             }
         }
 
@@ -124,6 +125,7 @@ namespace MaxMind.GeoIP2.UnitTests
             Assert.Equal(1221, response.AutonomousSystemNumber);
             Assert.Equal("Telstra Pty Ltd", response.AutonomousSystemOrganization);
             Assert.Equal(ipAddress, response.IPAddress);
+            Assert.Equal("1.128.0.0/11", response.Network.ToString());
         }
 
 
@@ -136,6 +138,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 var response = reader.ConnectionType(ipAddress);
                 Assert.Equal("Cable/DSL", response.ConnectionType);
                 Assert.Equal(ipAddress, response.IPAddress);
+                Assert.Equal("1.0.1.0/24", response.Network.ToString());
             }
         }
 
@@ -148,6 +151,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 var response = reader.Domain(ipAddress);
                 Assert.Equal("maxmind.com", response.Domain);
                 Assert.Equal(ipAddress, response.IPAddress);
+                Assert.Equal("1.2.0.0/16", response.Network.ToString());
             }
         }
 
@@ -168,6 +172,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 Assert.Equal("Cable/DSL", response.Traits.ConnectionType);
                 Assert.True(response.Traits.IsLegitimateProxy);
                 Assert.Equal(ipAddress, response.Traits.IPAddress);
+                Assert.Equal("74.209.16.0/20", response.Traits.Network.ToString());
             }
         }
 
@@ -183,6 +188,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 Assert.Equal("Telstra Internet", response.Isp);
                 Assert.Equal("Telstra Internet", response.Organization);
                 Assert.Equal(ipAddress, response.IPAddress);
+                Assert.Equal("1.128.0.0/11", response.Network.ToString());
             }
         }
 
@@ -196,6 +202,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 Assert.True(response.Country.IsInEuropeanUnion);
                 Assert.False(response.RegisteredCountry.IsInEuropeanUnion);
                 Assert.False(response.RepresentedCountry.IsInEuropeanUnion);
+                Assert.Equal("81.2.69.160/27", response.Traits.Network.ToString());
             }
         }
 
@@ -210,6 +217,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 Assert.Equal("US", response.RegisteredCountry.IsoCode);
                 Assert.False(response.RegisteredCountry.IsInEuropeanUnion);
                 Assert.False(response.RepresentedCountry.IsInEuropeanUnion);
+                Assert.Equal("81.2.69.160/27", response.Traits.Network.ToString());
             }
         }
 
@@ -224,6 +232,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 Assert.Equal(100, response.Location.AccuracyRadius);
                 Assert.False(response.RegisteredCountry.IsInEuropeanUnion);
                 Assert.False(response.RepresentedCountry.IsInEuropeanUnion);
+                Assert.Equal("81.2.69.160/27", response.Traits.Network.ToString());
             }
         }
 
@@ -238,6 +247,7 @@ namespace MaxMind.GeoIP2.UnitTests
                 Assert.Equal("London", response.City.Name);
                 Assert.False(response.RegisteredCountry.IsInEuropeanUnion);
                 Assert.False(response.RepresentedCountry.IsInEuropeanUnion);
+                Assert.Equal("81.2.69.160/27", response.Traits.Network.ToString());
             }
         }